+++ src/pkgdefaults.jl
@@ -22,7 +22,7 @@ const c² = Quantity(convert(Int64, ustrip(c0)), unit(c0))^2
 @eqrelation MassEnergy Energy/Mass = c²
 
 """
-    PhotonEnergy(; frequency=:linear, wavelength=:linear, wavenumber=:linear)
+    Spectral(; frequency=:linear, wavelength=:linear, wavenumber=:linear)
 
 Equivalence that relates the energy of a photon to its (linear or angular) frequency,
 wavelength, and wavenumber. Whether to convert to linear or angular quantities is
@@ -44,63 +44,63 @@ Equivalent quantities are converted according to the relations
 # Examples
 
 ```jldoctest
-julia> uconvert(u"nm", 13.6u"eV", PhotonEnergy()) # photon wavelength needed to ionize hydrogen
+julia> uconvert(u"nm", 13.6u"eV", Spectral()) # photon wavelength needed to ionize hydrogen
 91.16485178911785 nm
 
-julia> uconvert(u"Hz", 589u"nm", PhotonEnergy(frequency=:angular)) # angular frequency of sodium D line
+julia> uconvert(u"Hz", 589u"nm", Spectral(frequency=:angular)) # angular frequency of sodium D line
 3.1980501991661345e15 Hz
 ```
 """
-struct PhotonEnergy{freq, len, num} <: Equivalence
-    function PhotonEnergy{freq, len, num}() where {freq, len, num}
-        check_photonarg(freq)
-        check_photonarg(len)
-        check_photonarg(num)
+struct Spectral{freq, len, num} <: Equivalence
+    function Spectral{freq, len, num}() where {freq, len, num}
+        check_photonarg(freq, :frequency)
+        check_photonarg(len, :wavelength)
+        check_photonarg(num, :wavenumber)
         new()
     end
 end
 
-@inline check_photonarg(s::Symbol) =
-    s === :linear || s === :angular || throw(ArgumentError("PhotonEnergy parameter must be :linear or :angular"))
+@inline check_photonarg(s::Symbol, arg::Symbol) =
+    s === :linear || s === :angular || throw(ArgumentError("`$arg` argument must be :linear or :angular"))
 
-PhotonEnergy(; frequency=:linear, wavelength=:linear, wavenumber=:linear) =
-    PhotonEnergy{frequency, wavelength, wavenumber}()
+Spectral(; frequency=:linear, wavelength=:linear, wavenumber=:linear) =
+    Spectral{frequency, wavelength, wavenumber}()
 
-Base.show(io::IO, e::PhotonEnergy{freq, len, num}) where {freq, len, num} =
-    print(io, PhotonEnergy, "(frequency=", repr(freq), ", wavelength=", repr(len), ", wavenumber=", repr(num), ")")
+Base.show(io::IO, e::Spectral{freq, len, num}) where {freq, len, num} =
+    print(io, Spectral, "(frequency=", repr(freq), ", wavelength=", repr(len), ", wavenumber=", repr(num), ")")
 
-@eqrelation PhotonEnergy{:linear}  Energy/Frequency = h
-@eqrelation PhotonEnergy{:angular} Energy/Frequency = ħ
+@eqrelation Spectral{:linear}  Energy/Frequency = h
+@eqrelation Spectral{:angular} Energy/Frequency = ħ
 
-@eqrelation PhotonEnergy{F,:linear}  where F Energy*Length = h*c0
-@eqrelation PhotonEnergy{F,:angular} where F Energy*Length = ħ*c0
+@eqrelation Spectral{F,:linear}  where F Energy*Length = h*c0
+@eqrelation Spectral{F,:angular} where F Energy*Length = ħ*c0
 
-@eqrelation PhotonEnergy{F,L,:linear}  where {F,L} Energy/Wavenumber = h*c0
-@eqrelation PhotonEnergy{F,L,:angular} where {F,L} Energy/Wavenumber = ħ*c0
+@eqrelation Spectral{F,L,:linear}  where {F,L} Energy/Wavenumber = h*c0
+@eqrelation Spectral{F,L,:angular} where {F,L} Energy/Wavenumber = ħ*c0
 
-function edconvert(::dimtype(Frequency), x::Length, ::PhotonEnergy{F,L}) where {F,L}
+function edconvert(::dimtype(Frequency), x::Length, ::Spectral{F,L}) where {F,L}
     F === L       ? c0/x      :
     F === :linear ? c0/(π*2x) : 2c0*(π/x)
 end
-function edconvert(::dimtype(Length), x::Frequency, ::PhotonEnergy{F,L}) where {F,L}
+function edconvert(::dimtype(Length), x::Frequency, ::Spectral{F,L}) where {F,L}
     F === L       ? c0/x      :
     F === :linear ? c0/(π*2x) : 2c0*(π/x)
 end
 
-function edconvert(::dimtype(Frequency), x::Wavenumber, ::PhotonEnergy{F,L,N}) where {F,L,N}
+function edconvert(::dimtype(Frequency), x::Wavenumber, ::Spectral{F,L,N}) where {F,L,N}
     F === N       ? c0*x       :
     F === :linear ? c0/2*(x/π) : 2c0*(π*x)
 end
-function edconvert(::dimtype(Wavenumber), x::Frequency, ::PhotonEnergy{F,L,N}) where {F,L,N}
+function edconvert(::dimtype(Wavenumber), x::Frequency, ::Spectral{F,L,N}) where {F,L,N}
     F === N       ? x/c0       :
     F === :linear ? 2*(π*x)/c0 : (x/π)/2c0
 end
 
-function edconvert(::dimtype(Length), x::Wavenumber, ::PhotonEnergy{F,L,N}) where {F,L,N}
+function edconvert(::dimtype(Length), x::Wavenumber, ::Spectral{F,L,N}) where {F,L,N}
     L === N       ? inv(x)  :
     L === :linear ? 2*(π/x) : inv(2*(π*x))
 end
-function edconvert(::dimtype(Wavenumber), x::Length, ::PhotonEnergy{F,L,N}) where {F,L,N}
+function edconvert(::dimtype(Wavenumber), x::Length, ::Spectral{F,L,N}) where {F,L,N}
     L === N       ? inv(x)  :
     L === :linear ? 2*(π/x) : inv(2*(π*x))
 end